web_vuln_triage/server/requirements.txt
openenv-core>=0.1.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0